+++ providers/ic/src/utils/identity.utils.ts
@@ -1,18 +1,12 @@
 import {Identity} from '@dfinity/agent';
-import {LocalStorage} from '@dfinity/auth-client';
+import {IdbStorage} from '@dfinity/auth-client';
 import {DelegationChain, DelegationIdentity, Ed25519KeyIdentity} from '@dfinity/identity';
-import {InternetIdentityAuth} from '../types/identity';
 
-export const internetIdentityAuth = async (): Promise<InternetIdentityAuth> => {
-  const storage: LocalStorage = new LocalStorage('ic-');
-
-  const identityKey: string | null = await storage.get('identity');
-  const delegationChain: string | null = await storage.get('delegation');
-
-  return {
-    identityKey,
-    delegationChain
-  };
+export const internetIdentityAuth = async (): Promise<
+  [delegationChain: string | null, identityKey: string | null]
+> => {
+  const idbStorage: IdbStorage = new IdbStorage();
+  return Promise.all([idbStorage.get('delegation'), idbStorage.get('identity')]);
 };
 
 export const initIdentity = ({