+++ providers/ic/src/workers/idle.ic.worker.ts
@@ -1,22 +1,15 @@
 import {isDelegationValid} from '@dfinity/authentication';
 import {DelegationChain} from '@dfinity/identity';
-import type {InternetIdentityAuth} from '../types/identity';
 import {SignOutWindow} from '../types/sync.window';
+import {internetIdentityAuth} from '../utils/identity.utils';
 
 let idleTimer: boolean = false;
 
-export const startIdleTime = async (
-  {
-    internetIdentity
-  }: {
-    internetIdentity: InternetIdentityAuth;
-  },
-  onSignOut: SignOutWindow
-) => {
+export const startIdleTime = async (onSignOut: SignOutWindow) => {
   idleTimer = true;
 
   while (idleTimer) {
-    onIdleSignOut(internetIdentity, onSignOut);
+    await onIdleSignOut(onSignOut);
 
     // Sleep. setInterval not supported - throw an error upon trying to use postmessage "onSignOut" callback
     await new Promise((r) => setTimeout(r, 5000));
@@ -27,8 +20,8 @@ const stopTimer = () => (idleTimer = false);
 
 export const stopIdleTimer = async () => stopTimer();
 
-const onIdleSignOut = (internetIdentity: InternetIdentityAuth, onSignOut: SignOutWindow) => {
-  const {delegationChain} = internetIdentity;
+const onIdleSignOut = async (onSignOut: SignOutWindow) => {
+  const [delegationChain] = await internetIdentityAuth();
 
   if (delegationChain === null) {
     return;